Start work on a GC lowering pass
It won't be perfect, and I'm inexperienced in C++, but this pass should
eventually (hopefully) lower most Wasm GC instructions to linear memory.

Wasm GC technically depends on the function references proposal, but
lowering function references might be better suited for a separate pass,
since function references could potentially be lowered to a GC struct
containing a table index. Stringrefs could also be lowered in a similar
manner (a separate pass) to GC arrays. Such a combination could/would be
useful to not only AssemblyScript but also the WebAssembly ecosystem as
a whole. (Maybe Dart? Java? JVM bytecode? CIL? Not sure.)