Add support for i31ref
This pass now assumes that the allocation function always returns a
pointer with an alignment of at least 2, which would mean pointers to
structs and arrays would also have an LSB of 0. As a result, the LSB can
be (and now is) used as a flag to indicate a pointer represents an
i31ref, in which case the remaining 31 bits comprise its value.